refactor: isMyGoalAtom 위치 이동
