--- src/features/goal/atom/goalId.ts
@@ -1,3 +1,0 @@
-import { atom } from 'jotai';
-
-export const goalIdAtom = atom(-1);
