--- src/features/goal/atom/index.ts
@@ -1,2 +1,0 @@
-export { goalIdAtom } from './goalId';
-export { isMyGoalAtom } from './isMyGoal';
