+++ src/features/goal/atoms/index.ts
@@ -1,1 +1,2 @@
 export { goalIdAtom } from './goalId';
+export { isMyGoalAtom } from './isMyGoal';
