+++ src/features/goal/components/detail/GoalDetailContent.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import { Suspense, useEffect, useState } from 'react';
-import { useAtom, useSetAtom } from 'jotai';
+import { useSetAtom } from 'jotai';
 
 import { Skeleton } from '@/components';
 import { useGetGoal } from '@/hooks/reactQuery/goal';
 
-import { goalIdAtom, isMyGoalAtom } from '../../atom';
+import { goalIdAtom, isMyGoalAtom } from '../../atoms';
 
 import { AddTaskInput } from './AddTaskInput';
 import DetailLayout from './DetailLayout';
@@ -16,7 +16,7 @@ import { AddSubGoalPrompt, ContentBody, DetailFooterButton, DetailHeader, Sticke
 
 export const GoalDetailContent = ({ id }: { id: number }) => {
   const { data: goal } = useGetGoal({ goalId: Number(id) });
-  const [isMyGoal, setIsMyGoal] = useAtom(isMyGoalAtom);
+  const setIsMyGoal = useSetAtom(isMyGoalAtom);
   const setGoalId = useSetAtom(goalIdAtom);
   const [isOpenTaskInput, setOpenTaskInput] = useState(false);
 
@@ -47,7 +47,7 @@ export const GoalDetailContent = ({ id }: { id: number }) => {
               {goal.tasks.length ? (
                 <Tasks tasks={goal.tasks} onOpenInput={handleOpenTaskInput(true)} />
               ) : (
-                isMyGoal && !isOpenTaskInput && <AddSubGoalPrompt onClick={handleOpenTaskInput(true)} />
+                goal.isMyGoal && !isOpenTaskInput && <AddSubGoalPrompt onClick={handleOpenTaskInput(true)} />
               )}
             </div>
           )
